refactor: 移动行为函数到 `gui.ts`
