+++ src/constants/YumeSignEnum.ts
@@ -65,28 +65,3 @@ export enum BEHAVIOR_ZH {
     recycle = '回收', // item and exp
     disconnect = '销毁',
 }
-
-export const BEHAVIOR_FUNCTION = {
-    lookAtEntity: (sim: SimulatedPlayer, player: Player) => sim.lookAtEntity(player, LookDuration.Instant),
-    teleport: (sim: SimulatedPlayer, player: Player) => sim.teleport(player.location),
-    useAndStopUsingItem: (sim: SimulatedPlayer & Player) => sim.useItemInSlot(sim.selectedSlotIndex) && sim.stopUsingItem(),
-    useItemInSlot: (sim: SimulatedPlayer & Player) => sim.useItemInSlot(sim.selectedSlotIndex),
-    stopUsingItem: (sim: SimulatedPlayer) => sim.stopUsingItem(),
-    interact: (sim: SimulatedPlayer) => sim.interact(),
-    swapMainhandItem: (sim: SimulatedPlayer, player: Player) => commandManager.run('假人主手物品交换', { player, simulatedPlayer: sim }),
-    swapInventory: (sim: SimulatedPlayer, player: Player) => commandManager.run('假人背包交换', { player, simulatedPlayer: sim }),
-    swapEquipment: (sim: SimulatedPlayer, player: Player) => commandManager.run('假人装备交换', { player, simulatedPlayer: sim }),
-    rename: async (sim: SimulatedPlayer, player: Player) => {
-        const modalForm = new ModalFormData().title("假人改名");
-        modalForm.textField(`由 "${sim.nameTag}" 改为：`, '输入新名称', sim.nameTag);
-        const { canceled, formValues } = await modalForm.show(player);
-        if (canceled) return;
-
-        sim.nameTag = <string>formValues[0];
-        // commandManager.executeCommand('假人改名', [name], { entity: player, sim });
-    },
-    recycle: (sim: SimulatedPlayer, player: Player) => commandManager.run('假人资源回收', { player, simulatedPlayer: sim }), // item and exp
-    disconnect: (sim: SimulatedPlayer) => commandManager.run('假人销毁', { simulatedPlayer: sim }),
-};
-
-export const exeBehavior = (behavior: string) => BEHAVIOR[behavior] && BEHAVIOR_FUNCTION[behavior];
